% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form.R
\name{form}
\alias{form}
\alias{form.data.frame}
\alias{form.matrix}
\alias{form.table}
\title{Form Dataset}
\usage{
form(data, ...)

\method{form}{data.frame}(data, format = "\%.3f", formatcolumns = list(),
  marker = list(), symbols = list(), groups = list(),
  groups_only = FALSE, filterrows = character(),
  filtercolumns = character())

\method{form}{matrix}(data, ...)

\method{form}{table}(data, format = "\%i", ...)
}
\arguments{
\item{data}{A dataset with numeric values and other typed data.
Every column will be coerced to \code{character}.}

\item{format}{Default C-formatting (see \code{\link{sprintf}}) of all
numerical values.}

\item{formatcolumns}{A list with column-specific C-formatting of numerical
values. The format is identical to \code{format}. Regular
expressions (see \code{\link{regex}}) are used to select the
columns by column names. Example for two formattings:
\code{list("\%2.3f" = c("mean", "std"), "\%1.4f" =
"pvalue.*"))}}

\item{marker}{A list with marking rules which are going to be applied
to user-specified columns of \code{data}. Syntax:
\code{<rule> = <vektor of regular expressions>}. Possible rules include
column-wise comparing rules \code{min}, \code{max}; and arbitrary
element-wise comparing rules "\code{...v...}" where \code{v}
is a placeholder for the numeric value of a cell in a
logical expression. See Examples.}

\item{symbols}{A list with symbols to visually highlight all cells where the
marking rules of \code{marker} applied. Syntax:
\code{<symbol> = <vector of regular expressions for marking rules>}.
Arbitrary symbols are allowed.
Examples are "\code{***}", "\code{+}" or "\code{[v]}" and
"\code{\\mathbf{v}}" where \code{v} is a placeholder for the
(formatted) numeric value of a cell. The marker expressions that can
be approached by the regular expressions are \code{<rule>.<group>}
where \code{<group>} is one name specified in \code{groups}.
\code{<rule>.all} refers to the marks from \code{<rule>} applied
to the whole column.}

\item{groups}{A list with vectors of regular expressions that are going to be
applied to the row names of \code{data}. The specfied
column-wise \code{marker} are applied for each group. Syntax:
\code{<groupname> = <vector of regular expressions>}.
Do not name groups with "." (dot) or "all"!}

\item{groups_only}{A logical value indicating whether to apply the marker
rules only groupwise.}

\item{filterrows}{A vector of regular expressions which are applied to
the row names of \code{data} for selecting and sorting
the rows of the formatted table. Default: No filtering,
all rows are included as they appear.}

\item{filtercolumns}{A vector of regular expressions which are applied to
the column names of \code{data} for selecting and
sorting the columns of the formatted table. Default:
No filtering, all rows are included as they appear.}
}
\value{
The function returns a formatted character dataset of class
        \code{form} which inherits of \code{\link{data.frame}}.
}
\description{
Forms a dataset by applying flexible rules for numeric
             formatting, accentuation and filtering.
}
\details{
The function allows the user to format all kinds of numeric or
         character data on the fly, e.g. resulting output from computations.
         It can be employed greatly to automatically prepare data and results
         to review it directly in the R session or to export it to a
         dynamically designed LaTeX table code via \code{\link{latex}}.

         The package and this function will be improved by missing
         functionality on a regular basis. The author is eager to receive
         your suggestions in order to improve the dynamics and power of
         PlainT and \code{form}. Please write to Fabian Raters
         (mail@qrat.de).
}
\examples{
## forming mtcars
mtcarsf <- form(data = mtcars,
                format = "\%i",
                formatcolumns = list("\%5.1f" = "disp",
                                     "\%3i" = "hp",
                                     "\%4.1f" = "qsec",
                                     "\%4.2f" = c("drat", "wt")),
                marker = list(min = c("mpg", "cyl", "disp", "hp", "drat",
                                      "wt", "qsec"),
                              max = c("mpg", "cyl", "disp", "hp", "drat",
                                      "wt", "qsec"),
                              "v==1" = "am"),
                symbols = list("**v" = "max.all",
                               "..v" = "min.all",
                               "[v]" = "v==1"))
}
\seealso{
For exporting the formed frame, see \code{\link{latex}}.
}

