% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plaint.R
\docType{package}
\name{plaint}
\alias{plaint}
\alias{plaint-package}
\title{plaint: Plain Table Markup Language}
\description{
\code{plaint} allows you to ease your life with forming data arrays, such
as data frames, matrices or tables, and their export to user-designed
LaTeX tables. It comes with full S3 method support for further R objects.
Get back to the important parts of your work.
}
\details{
This packages provides two S3-type functions: \code{\link{form}} formats,
accentuates and filters data arrays and \code{\link{latex}} exports 'formed'
data frames to LaTeX code resulting in state of the art tables. Both
functions allow the user to easily transform arbitrary dataset based on a
user-specified set of rules and table markup. Read this documentation, the
vignettes and study the examples in order to see how you can employ these
functions with joy.
}

